+++ src/Attributes/Data.php
@@ -1,0 +1,20 @@
+<?php
+declare(strict_types=1);
+
+namespace MyTester\Attributes;
+
+use Attribute;
+
+/**
+ * Data attribute
+ * Defines one data set for a test, can be used multiple times
+ *
+ * @author Jakub Konečný
+ */
+#[Attribute(Attribute::TARGET_METHOD | Attribute::IS_REPEATABLE)]
+final readonly class Data
+{
+    public function __construct(public array $value)
+    {
+    }
+}