+++ tests/TestCaseTest.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace MyTester;
 
+use MyTester\Attributes\Data;
 use MyTester\Attributes\DataProvider as DataProviderAttribute;
 use MyTester\Attributes\DataProviderExternal;
 use MyTester\Attributes\IgnoreDeprecations;
@@ -89,6 +90,13 @@ final class TestCaseTest extends TestCase
     {
         $this->assertContains("a", $text);
         $this->assertGreaterThan(0, $number);
+    }
+
+    #[Data(["abc", ])]
+    #[Data(["abcd", ])]
+    public function testParamsData(string $text): void
+    {
+        $this->assertContains("a", $text);
     }
 
     public function dataProvider(): array
@@ -219,6 +227,7 @@ final class TestCaseTest extends TestCase
                 "testParamsMulti",
                 "testParamsIterator",
                 "testParamsExternal",
+                "testParamsData",
                 "testTestName",
                 "testSkip",
                 "testSkipPhpVersion",
@@ -249,7 +258,7 @@ final class TestCaseTest extends TestCase
     public function testGetJobs(): void
     {
         $jobs = $this->getJobs();
-        $this->assertCount(27, $jobs);
+        $this->assertCount(29, $jobs);
 
         $job = $jobs[0];
         $this->assertSame("TestCase::testState", $job->name);
@@ -351,6 +360,24 @@ final class TestCaseTest extends TestCase
         $this->assertTrue($job->reportDeprecations);
 
         $job = $jobs[11];
+        $this->assertSame("TestCase::testParamsData", $job->name);
+        $this->assertSame([$this, "testParamsData", ], $job->callback);
+        $this->assertSame(["abc", ], $job->params);
+        $this->assertFalse((bool) $job->skip);
+        $this->assertCount(1, $job->onAfterExecute);
+        $this->assertSame("", $job->dataSetName);
+        $this->assertTrue($job->reportDeprecations);
+
+        $job = $jobs[12];
+        $this->assertSame("TestCase::testParamsData", $job->name);
+        $this->assertSame([$this, "testParamsData", ], $job->callback);
+        $this->assertSame(["abcd", ], $job->params);
+        $this->assertFalse((bool) $job->skip);
+        $this->assertCount(1, $job->onAfterExecute);
+        $this->assertSame("", $job->dataSetName);
+        $this->assertTrue($job->reportDeprecations);
+
+        $job = $jobs[13];
         $this->assertSame("Custom name", $job->name);
         $this->assertSame([$this, "testTestName", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -359,7 +386,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[12];
+        $job = $jobs[14];
         $this->assertSame("Skip", $job->name);
         $this->assertSame([$this, "testSkip", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -368,7 +395,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[13];
+        $job = $jobs[15];
         $this->assertSame("PHP version", $job->name);
         $this->assertSame([$this, "testSkipPhpVersion", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -377,7 +404,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[14];
+        $job = $jobs[16];
         $this->assertSame("CGI sapi", $job->name);
         $this->assertSame([$this, "testCgiSapi", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -386,7 +413,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[15];
+        $job = $jobs[17];
         $this->assertSame("Extension", $job->name);
         $this->assertSame([$this, "testSkipExtension", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -395,7 +422,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[16];
+        $job = $jobs[18];
         $this->assertSame("OS family", $job->name);
         $this->assertSame([$this, "testSkipOsFamily", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -404,7 +431,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[17];
+        $job = $jobs[19];
         $this->assertSame("No assertions", $job->name);
         $this->assertSame([$this, "testNoAssertions", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -413,7 +440,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[18];
+        $job = $jobs[20];
         $this->assertSame("Deprecation", $job->name);
         $this->assertSame([$this, "testDeprecation", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -426,7 +453,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertFalse($job->reportDeprecations);
 
-        $job = $jobs[19];
+        $job = $jobs[21];
         $this->assertSame("TestCase::testGetSuiteName", $job->name);
         $this->assertSame([$this, "testGetSuiteName", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -435,7 +462,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[20];
+        $job = $jobs[22];
         $this->assertSame("TestCase::testGetJobName", $job->name);
         $this->assertSame([$this, "testGetJobName", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -444,7 +471,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[21];
+        $job = $jobs[23];
         $this->assertSame("TestCase::testGetTestMethodsNames", $job->name);
         $this->assertSame([$this, "testGetTestMethodsNames", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -453,7 +480,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[22];
+        $job = $jobs[24];
         $this->assertSame("TestCase::testShouldReportDeprecations", $job->name);
         $this->assertSame([$this, "testShouldReportDeprecations", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -462,7 +489,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[23];
+        $job = $jobs[25];
         $this->assertSame("TestCase::testGetJobs", $job->name);
         $this->assertSame([$this, "testGetJobs", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -471,7 +498,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[24];
+        $job = $jobs[26];
         $this->assertSame("TestCase::testIncomplete", $job->name);
         $this->assertSame([$this, "testIncomplete", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -480,7 +507,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[25];
+        $job = $jobs[27];
         $this->assertSame("TestCase::testSkipInside", $job->name);
         $this->assertSame([$this, "testSkipInside", ], $job->callback);
         $this->assertSame([], $job->params);
@@ -489,7 +516,7 @@ final class TestCaseTest extends TestCase
         $this->assertSame("", $job->dataSetName);
         $this->assertTrue($job->reportDeprecations);
 
-        $job = $jobs[26];
+        $job = $jobs[28];
         $this->assertSame("TestCase::testWhatever", $job->name);
         $this->assertSame([$this, "testWhatever", ], $job->callback);
         $this->assertSame([], $job->params);