+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Version 8.0.0-dev
 - added option to suppress warning about no performed assertions in a test method/test case
 - possible BC break: attribute Test, TestSuite and DataProvider are now readonly
 - added support for external data providers
+- BC break: added method getAnnotationMulti to interface IAnnotationsReaderEngine
+- added support for simple data sets with attribute Data
 
 Version 7.3.1
 - allowed installation konecnyjakub/event-dispatcher 2